Took out the ipfs saazuki API query
